refactor(tests): split tests/test_tools.py into tests/tools package
