--- tests/test_tools.py
@@ -1,1287 +1,0 @@
-import json
-from unittest.mock import MagicMock, patch
-
-import pytest
-
-from src.config import Settings
-from src.llms import ToolCall
-from src.tools import (
-    ToolContext,
-    ToolExecutor,
-    ToolRegistry,
-    default_registry,
-)
-from src.tools.slack import fetch_thread_history, read_attached_images
-from src.tools.image import generate_image
-from src.tools.search import search_web
-
-
-def _settings(**overrides) -> Settings:
-    base = {
-        "slack_bot_token": "xoxb-test",
-        "slack_signing_secret": "sig",
-        "llm_provider": "openai",
-        "llm_model": "gpt-4o-mini",
-        "image_provider": "openai",
-        "image_model": "gpt-image-1",
-        "agent_max_steps": 3,
-        "response_language": "ko",
-        "dynamodb_table_name": "t",
-        "aws_region": "us-east-1",
-    }
-    base.update(overrides)
-    return Settings(**base)
-
-
-def _ctx(event=None, slack_client=None, llm=None):
-    return ToolContext(
-        slack_client=slack_client or MagicMock(),
-        channel="C1",
-        thread_ts="ts1",
-        event=event or {},
-        settings=_settings(),
-        llm=llm or MagicMock(),
-    )
-
-
-# --------------------------------------------------------------------------- #
-# Registry
-# --------------------------------------------------------------------------- #
-
-
-def test_default_registry_has_expected_tools():
-    names = set(default_registry.names())
-    assert {
-        "read_attached_images",
-        "fetch_thread_history",
-        "search_web",
-        "generate_image",
-        "get_current_time",
-        "read_attached_document",
-    }.issubset(names)
-    assert "search_slack_messages" not in names  # removed — user-token only, tied to installer
-
-
-def test_registry_specs_match_llm_shape():
-    for spec in default_registry.specs():
-        assert set(spec.keys()) == {"name", "description", "parameters"}
-        assert spec["parameters"]["type"] == "object"
-
-
-# --------------------------------------------------------------------------- #
-# Executor
-# --------------------------------------------------------------------------- #
-
-
-def test_executor_unknown_tool():
-    registry = ToolRegistry()
-    executor = ToolExecutor(_ctx(), registry)
-    result = executor.execute(ToolCall(id="1", name="nope", arguments={}))
-    assert result["ok"] is False
-    assert "unknown tool" in result["error"]
-
-
-def test_executor_timeout_guards_slow_tools():
-    import time
-
-    registry = ToolRegistry()
-
-    def slow(ctx):
-        time.sleep(1.0)
-
-    from src.tools.registry import ToolDef
-
-    registry.register(ToolDef(name="slow", description="", parameters={"type": "object", "properties": {}}, fn=slow))
-    executor = ToolExecutor(_ctx(), registry, timeout=0.1)
-    result = executor.execute(ToolCall(id="1", name="slow", arguments={}))
-    assert result["ok"] is False
-    assert "timed out" in result["error"]
-
-
-def test_executor_wraps_boto_client_error():
-    """Bedrock invoke failures (botocore ClientError) must be returned as
-    {ok: False, error: ...} so the LLM can plan around the failure instead
-    of the exception bubbling out of the agent loop."""
-    from botocore.exceptions import ClientError
-
-    registry = ToolRegistry()
-
-    def failing_bedrock(ctx):
-        raise ClientError(
-            {"Error": {"Code": "ResourceNotFoundException", "Message": "Legacy model"}},
-            "InvokeModel",
-        )
-
-    from src.tools.registry import ToolDef
-
-    registry.register(
-        ToolDef(
-            name="bedrock_thing",
-            description="",
-            parameters={"type": "object", "properties": {}},
-            fn=failing_bedrock,
-        )
-    )
-    executor = ToolExecutor(_ctx(), registry)
-    result = executor.execute(ToolCall(id="1", name="bedrock_thing", arguments={}))
-    assert result["ok"] is False
-    assert "ResourceNotFoundException" in result["error"] or "Legacy" in result["error"]
-
-
-def test_executor_captures_tool_error():
-    registry = ToolRegistry()
-
-    def boom(ctx):
-        raise ValueError("nope")
-
-    from src.tools.registry import ToolDef
-
-    registry.register(ToolDef(name="boom", description="", parameters={"type": "object", "properties": {}}, fn=boom))
-    executor = ToolExecutor(_ctx(), registry)
-    result = executor.execute(ToolCall(id="1", name="boom", arguments={}))
-    assert result["ok"] is False
-    assert "nope" in result["error"]
-
-
-def test_executor_per_tool_timeout_override():
-    """A tool registered with its own timeout overrides the executor default."""
-    import time
-
-    registry = ToolRegistry()
-
-    def moderately_slow(ctx):
-        time.sleep(0.3)
-        return "done"
-
-    from src.tools.registry import ToolDef
-
-    registry.register(
-        ToolDef(
-            name="slowish",
-            description="",
-            parameters={"type": "object", "properties": {}},
-            fn=moderately_slow,
-            timeout=1.0,
-        )
-    )
-    # Default timeout short enough to kill a naïve tool; per-tool override lets
-    # this one finish.
-    executor = ToolExecutor(_ctx(), registry, timeout=0.1)
-    result = executor.execute(ToolCall(id="1", name="slowish", arguments={}))
-    assert result["ok"] is True
-    assert result["result"] == "done"
-
-
-def test_generate_image_tool_has_extended_timeout():
-    """Image generation is slow; its registered timeout must be > default."""
-    td = default_registry.get("generate_image")
-    assert td is not None
-    assert td.timeout is not None
-    assert td.timeout >= 60.0
-
-
-# --------------------------------------------------------------------------- #
-# read_attached_images SSRF guard
-# --------------------------------------------------------------------------- #
-
-
-def test_read_attached_images_rejects_non_slack_host():
-    event = {"files": [{"mimetype": "image/png", "url_private_download": "https://evil.example.com/x.png"}]}
-    with pytest.raises(ValueError):
-        read_attached_images(_ctx(event=event), limit=1)
-
-
-def test_read_attached_images_rejects_http_scheme():
-    event = {"files": [{"mimetype": "image/png", "url_private_download": "http://files.slack.com/x.png"}]}
-    with pytest.raises(ValueError):
-        read_attached_images(_ctx(event=event), limit=1)
-
-
-def test_read_attached_images_accepts_slack_host_variants():
-    event = {
-        "files": [
-            {"mimetype": "image/png", "url_private_download": "https://files-pri.slack.com/x.png", "name": "a"},
-        ]
-    }
-    llm = MagicMock()
-    llm.describe_image.return_value = "a cat"
-    ctx = _ctx(event=event, llm=llm)
-    with patch("src.tools.slack.urllib.request.urlopen") as opener:
-        opener.return_value.__enter__.return_value.read.return_value = b"fake"
-        result = read_attached_images(ctx, limit=1)
-    assert result == [{"name": "a", "summary": "a cat"}]
-
-
-def test_read_attached_images_skips_non_image_mimetypes():
-    event = {"files": [{"mimetype": "application/pdf", "url_private_download": "https://files.slack.com/x.pdf"}]}
-    assert read_attached_images(_ctx(event=event), limit=1) == []
-
-
-# --------------------------------------------------------------------------- #
-# fetch_thread_history
-# --------------------------------------------------------------------------- #
-
-
-def test_fetch_thread_history_resolves_user_files_and_reactions():
-    """History should carry display names, file metadata, and reactions so the
-    LLM can answer things like "누가 좋아요 눌렀어?" or "아까 그 이미지 분석해줘"."""
-    from src.slack_helpers import user_name_cache
-
-    # Reset the module-level cache so prior tests don't leak.
-    user_name_cache._cache.clear()
-
-    client = MagicMock()
-    client.conversations_replies.return_value = {
-        "messages": [
-            {
-                "user": "U1",
-                "text": "look at this",
-                "ts": "1713.1",
-                "files": [
-                    {
-                        "name": "cat.png",
-                        "mimetype": "image/png",
-                        "url_private_download": "https://files.slack.com/x/cat.png",
-                        "permalink": "https://slack/p1",
-                        "title": "cute",
-                    }
-                ],
-            },
-            {
-                "user": "U2",
-                "text": "nice!",
-                "ts": "1713.2",
-                "reactions": [
-                    {"name": "thumbsup", "count": 2, "users": ["U1", "U3"]},
-                ],
-            },
-        ]
-    }
-
-    def _users_info(user):
-        return {"user": {"profile": {"display_name": f"name-{user}"}}}
-
-    client.users_info.side_effect = _users_info
-
-    out = fetch_thread_history(_ctx(slack_client=client), limit=5)
-    assert len(out) == 2
-    first, second = out
-    assert first["user"] == "name-U1"
-    assert first["text"] == "look at this"
-    assert first["ts"] == "1713.1"
-    assert first["files"] == [
-        {
-            "name": "cat.png",
-            "mimetype": "image/png",
-            "url_private_download": "https://files.slack.com/x/cat.png",
-            "permalink": "https://slack/p1",
-            "title": "cute",
-        }
-    ]
-    assert first["reactions"] == []
-
-    assert second["user"] == "name-U2"
-    assert second["files"] == []
-    assert second["reactions"] == [
-        {"emoji": "thumbsup", "count": 2, "users": ["name-U1", "name-U3"]}
-    ]
-
-
-def test_read_attached_images_accepts_extra_urls():
-    """Images referenced from fetch_thread_history (url_private_download) must
-    be loadable via read_attached_images(urls=[...])."""
-    ctx = _ctx()
-    ctx.llm.describe_image.return_value = "a cat history"
-    with patch("src.tools.slack.urllib.request.urlopen") as opener:
-        opener.return_value.__enter__.return_value.read.return_value = b"fake-bytes"
-        out = read_attached_images(
-            ctx,
-            limit=5,
-            urls=["https://files.slack.com/x/cat.png"],
-        )
-    assert out == [{"name": "cat.png", "summary": "a cat history"}]
-
-
-def test_read_attached_images_urls_reject_non_slack_host():
-    ctx = _ctx()
-    with pytest.raises(ValueError):
-        read_attached_images(ctx, urls=["https://evil.example.com/cat.png"])
-
-
-def test_read_attached_images_respects_total_limit_across_event_and_urls():
-    event = {
-        "files": [
-            {
-                "mimetype": "image/png",
-                "url_private_download": "https://files.slack.com/e1.png",
-                "name": "e1.png",
-            }
-        ]
-    }
-    ctx = _ctx(event=event)
-    ctx.llm.describe_image.return_value = "desc"
-    with patch("src.tools.slack.urllib.request.urlopen") as opener:
-        opener.return_value.__enter__.return_value.read.return_value = b"x"
-        out = read_attached_images(
-            ctx,
-            limit=2,
-            urls=[
-                "https://files.slack.com/u1.png",
-                "https://files.slack.com/u2.png",  # should be skipped (limit=2)
-            ],
-        )
-    assert len(out) == 2
-    assert {item["name"] for item in out} == {"e1.png", "u1.png"}
-
-
-# --------------------------------------------------------------------------- #
-# search_web
-# --------------------------------------------------------------------------- #
-
-
-def test_search_web_ddg_parses_results():
-    ctx = _ctx()
-    payload = {
-        "AbstractURL": "https://example.com/a",
-        "AbstractText": "abstract",
-        "RelatedTopics": [{"Text": "t1", "FirstURL": "https://example.com/1"}],
-    }
-    with patch("src.tools.search.urllib.request.urlopen") as opener:
-        opener.return_value.__enter__.return_value.read.return_value = json.dumps(payload).encode()
-        results = search_web(ctx, query="q", limit=5)
-    assert results[0]["url"] == "https://example.com/a"
-    assert results[1]["url"] == "https://example.com/1"
-
-
-def test_search_web_uses_tavily_when_key_set():
-    ctx = ToolContext(
-        slack_client=MagicMock(),
-        channel="C1",
-        thread_ts="ts1",
-        event={},
-        settings=_settings(tavily_api_key="tvly-xyz"),
-        llm=MagicMock(),
-    )
-    payload = {"results": [{"title": "t", "url": "https://x", "content": "c"}]}
-    with patch("src.tools.search.urllib.request.urlopen") as opener:
-        opener.return_value.__enter__.return_value.read.return_value = json.dumps(payload).encode()
-        out = search_web(ctx, query="q", limit=5)
-    assert out == [{"title": "t", "url": "https://x", "content": "c"}]
-
-
-# --------------------------------------------------------------------------- #
-# generate_image
-# --------------------------------------------------------------------------- #
-
-
-def test_generate_image_returns_permalink():
-    llm = MagicMock()
-    llm.generate_image.return_value = b"imgbytes"
-    client = MagicMock()
-    client.files_upload_v2.return_value = {"file": {"permalink": "https://slack/abc", "title": "t"}}
-    ctx = _ctx(slack_client=client, llm=llm)
-    out = generate_image(ctx, prompt="cat")
-    assert out == {"permalink": "https://slack/abc", "title": "t"}
-    llm.generate_image.assert_called_once_with("cat")
-
-
-# --------------------------------------------------------------------------- #
-# get_current_time
-# --------------------------------------------------------------------------- #
-
-
-def test_get_current_time_uses_default_timezone():
-    from src.tools.time import get_current_time
-
-    ctx = _ctx()  # _settings() default_timezone defaults to Asia/Seoul
-    out = get_current_time(ctx)
-    assert out["timezone"] == "Asia/Seoul"
-    assert out["iso"].endswith("+09:00")
-    # Weekday is a full English day name (Monday..Sunday)
-    assert out["weekday"] in {
-        "Monday", "Tuesday", "Wednesday", "Thursday",
-        "Friday", "Saturday", "Sunday",
-    }
-    assert isinstance(out["unix"], int)
-
-
-def test_get_current_time_respects_custom_timezone():
-    from src.tools.time import get_current_time
-
-    ctx = _ctx()
-    out = get_current_time(ctx, timezone="UTC")
-    assert out["timezone"] == "UTC"
-    assert out["iso"].endswith("+00:00")
-
-
-def test_get_current_time_invalid_tz_via_executor():
-    """Invalid timezone should surface as {ok: False, error: ...} via the
-    executor so the LLM can recover."""
-    from src.tools import default_registry
-
-    executor = ToolExecutor(_ctx(), default_registry)
-    result = executor.execute(
-        ToolCall(id="t1", name="get_current_time", arguments={"timezone": "Narnia/Center"})
-    )
-    assert result["ok"] is False
-    assert "unknown timezone" in result["error"]
-
-
-# --------------------------------------------------------------------------- #
-# read_attached_document
-# --------------------------------------------------------------------------- #
-
-
-def test_read_attached_document_text_file():
-    from src.tools.slack import read_attached_document
-
-    event = {
-        "files": [
-            {
-                "mimetype": "text/plain",
-                "url_private_download": "https://files.slack.com/notes.txt",
-                "name": "notes.txt",
-            }
-        ]
-    }
-    ctx = _ctx(event=event)
-    body = b"Hello\n  world.\nLine 3."
-    with patch("src.tools.slack.urllib.request.urlopen") as opener:
-        resp = opener.return_value.__enter__.return_value
-        resp.read.return_value = body
-        resp.headers = {"Content-Length": str(len(body))}
-        out = read_attached_document(ctx, limit=1)
-    assert len(out) == 1
-    entry = out[0]
-    assert entry["name"] == "notes.txt"
-    assert entry["mimetype"] == "text/plain"
-    assert entry["truncated"] is False
-    assert "Hello" in entry["text"]
-    assert entry["chars"] == len(entry["text"])
-    assert entry["pages"] == 0  # text files report 0 pages
-
-
-def _build_pdf_bytes(pages_text: list[str]) -> bytes:
-    """Build a minimal PDF (one page per string) using reportlab. Test-only."""
-    from io import BytesIO
-    from reportlab.pdfgen.canvas import Canvas
-    from reportlab.lib.pagesizes import letter
-
-    buf = BytesIO()
-    canvas = Canvas(buf, pagesize=letter)
-    for text in pages_text:
-        canvas.drawString(72, 720, text)
-        canvas.showPage()
-    canvas.save()
-    return buf.getvalue()
-
-
-def _mock_pdf_response(opener, body: bytes, headers=None):
-    """Wire the urlopen mock to stream `body` in chunks through `_fetch_slack_file`."""
-    resp = opener.return_value.__enter__.return_value
-    buf = {"pos": 0}
-
-    def _chunked(n=-1):
-        if n == -1:
-            remaining = body[buf["pos"]:]
-            buf["pos"] = len(body)
-            return remaining
-        chunk = body[buf["pos"]:buf["pos"] + n]
-        buf["pos"] += len(chunk)
-        return chunk
-
-    resp.read.side_effect = _chunked
-    resp.headers = dict(headers or {"Content-Length": str(len(body)), "Content-Type": "application/pdf"})
-
-
-def test_read_attached_document_pdf_happy_path():
-    from src.tools.slack import read_attached_document
-
-    pdf = _build_pdf_bytes(["Hello PDF page one.", "Page two here."])
-    event = {
-        "files": [
-            {
-                "mimetype": "application/pdf",
-                "url_private_download": "https://files.slack.com/report.pdf",
-                "name": "report.pdf",
-            }
-        ]
-    }
-    ctx = _ctx(event=event)
-    with patch("src.tools.slack.urllib.request.urlopen") as opener:
-        _mock_pdf_response(opener, pdf)
-        out = read_attached_document(ctx, limit=1)
-    assert len(out) == 1
-    entry = out[0]
-    assert entry["name"] == "report.pdf"
-    assert entry["pages"] == 2
-    assert entry["truncated"] is False
-    assert entry["chars"] > 0
-
-
-def test_read_attached_document_pdf_truncation():
-    from src.tools.slack import read_attached_document
-
-    pdf = _build_pdf_bytes(["A" * 500])
-    event = {
-        "files": [
-            {
-                "mimetype": "application/pdf",
-                "url_private_download": "https://files.slack.com/big.pdf",
-                "name": "big.pdf",
-            }
-        ]
-    }
-    ctx = _ctx(
-        event=event,
-    )
-    ctx = ToolContext(
-        slack_client=ctx.slack_client,
-        channel=ctx.channel,
-        thread_ts=ctx.thread_ts,
-        event=ctx.event,
-        settings=_settings(max_doc_chars=50),
-        llm=ctx.llm,
-    )
-    with patch("src.tools.slack.urllib.request.urlopen") as opener:
-        _mock_pdf_response(opener, pdf)
-        out = read_attached_document(ctx, limit=1)
-    assert out[0]["truncated"] is True
-    assert out[0]["chars"] == 50
-
-
-def test_read_attached_document_page_cap():
-    from src.tools.slack import read_attached_document
-
-    pdf = _build_pdf_bytes(["p1", "p2", "p3"])
-    event = {
-        "files": [
-            {
-                "mimetype": "application/pdf",
-                "url_private_download": "https://files.slack.com/pages.pdf",
-                "name": "pages.pdf",
-            }
-        ]
-    }
-    ctx = ToolContext(
-        slack_client=MagicMock(),
-        channel="C1",
-        thread_ts="ts1",
-        event=event,
-        settings=_settings(max_doc_pages=2),
-        llm=MagicMock(),
-    )
-    with patch("src.tools.slack.urllib.request.urlopen") as opener:
-        _mock_pdf_response(opener, pdf)
-        out = read_attached_document(ctx, limit=1)
-    assert "error" in out[0]
-    assert "MAX_DOC_PAGES" in out[0]["error"]
-
-
-def test_read_attached_document_size_cap_via_content_length():
-    from src.tools.slack import read_attached_document
-
-    event = {
-        "files": [
-            {
-                "mimetype": "text/plain",
-                "url_private_download": "https://files.slack.com/huge.txt",
-                "name": "huge.txt",
-            }
-        ]
-    }
-    ctx = ToolContext(
-        slack_client=MagicMock(),
-        channel="C1",
-        thread_ts="ts1",
-        event=event,
-        settings=_settings(max_doc_bytes=100),  # tiny cap
-        llm=MagicMock(),
-    )
-    with patch("src.tools.slack.urllib.request.urlopen") as opener:
-        resp = opener.return_value.__enter__.return_value
-        resp.headers = {"Content-Length": "200"}  # > cap
-        resp.read.return_value = b"x" * 10  # should never be read past cap
-        out = read_attached_document(ctx, limit=1)
-    assert "error" in out[0]
-    assert "MAX_DOC_BYTES" in out[0]["error"]
-
-
-def test_read_attached_document_size_cap_via_streamed_read():
-    from src.tools.slack import read_attached_document
-
-    event = {
-        "files": [
-            {
-                "mimetype": "text/plain",
-                "url_private_download": "https://files.slack.com/nohead.txt",
-                "name": "nohead.txt",
-            }
-        ]
-    }
-    ctx = ToolContext(
-        slack_client=MagicMock(),
-        channel="C1",
-        thread_ts="ts1",
-        event=event,
-        settings=_settings(max_doc_bytes=100),
-        llm=MagicMock(),
-    )
-    body = b"y" * 200
-    with patch("src.tools.slack.urllib.request.urlopen") as opener:
-        resp = opener.return_value.__enter__.return_value
-        resp.headers = {}  # no Content-Length
-        buf = {"pos": 0}
-
-        def _chunked(n=-1):
-            if n == -1:
-                remaining = body[buf["pos"]:]
-                buf["pos"] = len(body)
-                return remaining
-            chunk = body[buf["pos"]:buf["pos"] + n]
-            buf["pos"] += len(chunk)
-            return chunk
-
-        resp.read.side_effect = _chunked
-        out = read_attached_document(ctx, limit=1)
-    assert "error" in out[0]
-    assert "MAX_DOC_BYTES" in out[0]["error"]
-
-
-def test_read_attached_document_rejects_non_slack_host():
-    from src.tools.slack import read_attached_document
-
-    ctx = _ctx()
-    out = read_attached_document(
-        ctx, urls=["https://evil.example.com/foo.pdf"], limit=1
-    )
-    assert len(out) == 1
-    assert "error" in out[0]
-    assert "invalid" in out[0]["error"].lower()
-
-
-def test_read_attached_document_skips_encrypted_pdf():
-    from src.tools.slack import read_attached_document
-    from io import BytesIO
-    from pypdf import PdfWriter
-
-    writer = PdfWriter()
-    writer.add_blank_page(width=612, height=792)
-    # NOTE: pypdf>=4.0 uses keyword-only user_password. If requirements.txt's
-    # upper pin is ever relaxed past 6.0, verify this signature still holds.
-    writer.encrypt(user_password="secret")
-    buf = BytesIO()
-    writer.write(buf)
-    encrypted_pdf = buf.getvalue()
-
-    event = {
-        "files": [
-            {
-                "mimetype": "application/pdf",
-                "url_private_download": "https://files.slack.com/enc.pdf",
-                "name": "enc.pdf",
-            }
-        ]
-    }
-    ctx = _ctx(event=event)
-    with patch("src.tools.slack.urllib.request.urlopen") as opener:
-        _mock_pdf_response(opener, encrypted_pdf)
-        out = read_attached_document(ctx, limit=1)
-    assert "error" in out[0]
-    assert "encrypted" in out[0]["error"]
-
-
-def test_read_attached_document_skips_image_mime():
-    from src.tools.slack import read_attached_document
-
-    event = {
-        "files": [
-            {
-                "mimetype": "image/png",
-                "url_private_download": "https://files.slack.com/a.png",
-                "name": "a.png",
-            }
-        ]
-    }
-    ctx = _ctx(event=event)
-    # urlopen should NOT be called — image MIMEs are filtered before fetch
-    with patch("src.tools.slack.urllib.request.urlopen") as opener:
-        out = read_attached_document(ctx, limit=1)
-    opener.assert_not_called()
-    assert out == []
-
-
-def test_read_attached_document_http_error_returns_per_item():
-    from src.tools.slack import read_attached_document
-    import urllib.error
-
-    event = {
-        "files": [
-            {
-                "mimetype": "application/pdf",
-                "url_private_download": "https://files.slack.com/missing.pdf",
-                "name": "missing.pdf",
-            }
-        ]
-    }
-    ctx = _ctx(event=event)
-    with patch("src.tools.slack.urllib.request.urlopen") as opener:
-        opener.side_effect = urllib.error.HTTPError(
-            url="https://files.slack.com/missing.pdf",
-            code=404,
-            msg="Not Found",
-            hdrs=None,
-            fp=None,
-        )
-        out = read_attached_document(ctx, limit=1)
-    assert len(out) == 1
-    assert "error" in out[0]
-    assert "404" in out[0]["error"]
-
-
-# --------------------------------------------------------------------------- #
-# fetch_webpage — SSRF guard
-# --------------------------------------------------------------------------- #
-
-
-def test_validate_public_https_url_rejects_http_scheme():
-    from src.tools.web import _validate_public_https_url
-
-    with pytest.raises(ValueError, match="https"):
-        _validate_public_https_url("http://example.com/")
-
-
-def test_validate_public_https_url_rejects_ip_literal_v4():
-    from src.tools.web import _validate_public_https_url
-
-    with pytest.raises(ValueError, match="IP literals"):
-        _validate_public_https_url("https://127.0.0.1/")
-
-
-def test_validate_public_https_url_rejects_ip_literal_v6():
-    from src.tools.web import _validate_public_https_url
-
-    with pytest.raises(ValueError, match="IP literals"):
-        _validate_public_https_url("https://[::1]/")
-
-
-def test_validate_public_https_url_rejects_private_dns(monkeypatch):
-    from src.tools.web import _validate_public_https_url
-
-    def fake_getaddrinfo(host, port, family=0, type=0, *args, **kwargs):
-        # Simulate DNS pointing at RFC1918 space.
-        return [(None, None, None, "", ("10.0.0.1", port))]
-
-    monkeypatch.setattr("src.tools.web.socket.getaddrinfo", fake_getaddrinfo)
-    with pytest.raises(ValueError, match="non-public"):
-        _validate_public_https_url("https://internal.corp.example/")
-
-
-def test_validate_public_https_url_rejects_metadata_host(monkeypatch):
-    from src.tools.web import _validate_public_https_url
-
-    def fake_getaddrinfo(host, port, family=0, type=0, *args, **kwargs):
-        return [(None, None, None, "", ("169.254.169.254", port))]
-
-    monkeypatch.setattr("src.tools.web.socket.getaddrinfo", fake_getaddrinfo)
-    with pytest.raises(ValueError, match="non-public"):
-        _validate_public_https_url("https://cloud.metadata.example/")
-
-
-def test_validate_public_https_url_accepts_public_host(monkeypatch):
-    from src.tools.web import _validate_public_https_url
-
-    def fake_getaddrinfo(host, port, family=0, type=0, *args, **kwargs):
-        return [(None, None, None, "", ("93.184.216.34", port))]  # example.com
-
-    monkeypatch.setattr("src.tools.web.socket.getaddrinfo", fake_getaddrinfo)
-    scheme, host = _validate_public_https_url("https://example.com/path")
-    assert scheme == "https"
-    assert host == "example.com"
-
-
-def test_validate_public_https_url_dns_failure(monkeypatch):
-    import socket as _socket
-
-    from src.tools.web import _validate_public_https_url
-
-    def fake_getaddrinfo(host, port, family=0, type=0, *args, **kwargs):
-        raise _socket.gaierror("nodename nor servname provided")
-
-    monkeypatch.setattr("src.tools.web.socket.getaddrinfo", fake_getaddrinfo)
-    with pytest.raises(ValueError, match="DNS resolution failed"):
-        _validate_public_https_url("https://nonexistent.invalid.example/")
-
-
-def test_no_redirect_handler_raises_on_302():
-    import urllib.error
-    import urllib.request
-
-    from src.tools.web import _NoRedirectHandler
-
-    handler = _NoRedirectHandler()
-    req = urllib.request.Request("https://example.com/")
-    with pytest.raises(urllib.error.HTTPError, match="redirects not allowed"):
-        handler.redirect_request(req, None, 302, "Found", {}, "https://evil.example/")
-
-
-# --------------------------------------------------------------------------- #
-# fetch_webpage — HTML parser + Jina response parser
-# --------------------------------------------------------------------------- #
-
-
-def test_html_text_extractor_basic():
-    from src.tools.web import _HtmlTextExtractor
-
-    html = (
-        "<html><head><title>  Hello </title></head>"
-        "<body>"
-        "<script>alert('x')</script>"
-        "<style>.a{}</style>"
-        "<h1>Heading</h1>"
-        "<p>Para one.</p>"
-        "<p>Para two.</p>"
-        "<a href='https://a.example/1'>Link A</a>"
-        "<a href='/relative'>Rel</a>"
-        "<a href='#frag'>Frag</a>"
-        "<a href='mailto:x@y'>Mail</a>"
-        "</body></html>"
-    )
-    x = _HtmlTextExtractor("https://base.example/page")
-    x.feed(html)
-    assert x.title() == "Hello"
-    text = x.text()
-    assert "Heading" in text
-    assert "Para one." in text
-    assert "Para two." in text
-    assert "alert" not in text
-    assert ".a{}" not in text
-    assert ("Link A", "https://a.example/1") in x.links
-    # Relative links resolved against base
-    assert ("Rel", "https://base.example/relative") in x.links
-    # mailto / fragment retained raw — filtering happens in _filter_links
-    assert any(url.startswith("mailto:") for _, url in x.links)
-
-
-def test_filter_links_drops_non_https_and_dedups():
-    from src.tools.web import _filter_links
-
-    raw = [
-        ("A", "https://a.example/1"),
-        ("A dup", "https://a.example/1#top"),  # dedups by fragment-stripped url
-        ("B", "https://b.example/"),
-        ("Self", "https://base.example/page"),  # self-ref dropped
-        ("Mail", "mailto:x@y"),
-        ("JS", "javascript:void(0)"),
-        ("HTTP", "http://insecure.example/"),
-    ]
-    out = _filter_links(raw, base_url="https://base.example/page", limit=10)
-    urls = [item["url"] for item in out]
-    assert urls == ["https://a.example/1", "https://b.example/"]
-    assert out[0]["title"] == "A"
-
-
-def test_filter_links_respects_limit():
-    from src.tools.web import _filter_links
-
-    raw = [(f"T{i}", f"https://x.example/{i}") for i in range(5)]
-    out = _filter_links(raw, base_url="https://base.example/", limit=3)
-    assert len(out) == 3
-    assert [item["url"] for item in out] == [
-        "https://x.example/0",
-        "https://x.example/1",
-        "https://x.example/2",
-    ]
-
-
-def test_extract_markdown_links_parses_inline_markdown():
-    from src.tools.web import _extract_markdown_links
-
-    md = (
-        "Title\n\nSome prose with [Google](https://google.com/about) "
-        "and [Self](https://base.example/page) and [Same](https://google.com/about?ref=x).\n"
-        "[Another](https://example.org/)"
-    )
-    out = _extract_markdown_links(md, base_url="https://base.example/page", limit=10)
-    urls = [item["url"] for item in out]
-    assert "https://google.com/about" in urls
-    assert "https://example.org/" in urls
-    assert "https://base.example/page" not in urls  # self-ref dropped
-
-
-def test_parse_jina_response_strips_header():
-    from src.tools.web import _parse_jina_response
-
-    payload = (
-        "Title: Example Page\n"
-        "URL Source: https://example.com/\n"
-        "Markdown Content:\n"
-        "# Heading\n\nBody text.\n"
-    )
-    title, body = _parse_jina_response(payload)
-    assert title == "Example Page"
-    assert body.startswith("# Heading")
-    assert "URL Source" not in body
-
-
-def test_parse_jina_response_no_header():
-    from src.tools.web import _parse_jina_response
-
-    payload = "just raw markdown\n\nwith no prefix"
-    title, body = _parse_jina_response(payload)
-    assert title == ""
-    assert body == payload
-
-
-def test_html_text_extractor_empty():
-    from src.tools.web import _HtmlTextExtractor
-
-    x = _HtmlTextExtractor("https://base.example/")
-    x.feed("")
-    assert x.title() == ""
-    assert x.text() == ""
-    assert x.links == []
-
-
-def test_filter_links_host_case_normalization():
-    from src.tools.web import _filter_links
-
-    out = _filter_links(
-        [
-            ("A", "https://Example.COM/path"),
-            ("B", "https://example.com/path"),   # dup by host case
-            ("Self", "https://BASE.EXAMPLE/page"),  # self-ref by host case
-        ],
-        base_url="https://base.example/page",
-        limit=10,
-    )
-    urls = [item["url"] for item in out]
-    assert urls == ["https://Example.COM/path"]  # first-seen wins, case preserved in output
-
-
-def test_extract_markdown_links_skips_images():
-    from src.tools.web import _extract_markdown_links
-
-    md = (
-        "![logo](https://img.example/logo.png) "
-        "see [Docs](https://docs.example/) for more"
-    )
-    out = _extract_markdown_links(md, base_url="https://base.example/", limit=10)
-    urls = [item["url"] for item in out]
-    assert "https://img.example/logo.png" not in urls
-    assert "https://docs.example/" in urls
-
-
-def test_extract_markdown_links_preserves_paren_in_url():
-    from src.tools.web import _extract_markdown_links
-
-    md = "see [Wiki](https://en.wikipedia.org/wiki/Foo_(bar)) for context"
-    out = _extract_markdown_links(md, base_url="https://base.example/", limit=10)
-    urls = [item["url"] for item in out]
-    assert "https://en.wikipedia.org/wiki/Foo_(bar)" in urls
-
-
-def test_parse_jina_response_inline_markdown_content():
-    from src.tools.web import _parse_jina_response
-
-    payload = "Title: T\nURL Source: https://example.com/\nMarkdown Content: # Heading\n\nBody here."
-    title, body = _parse_jina_response(payload)
-    assert title == "T"
-    assert body.startswith("# Heading")
-    assert "Body here." in body
-
-
-# --------------------------------------------------------------------------- #
-# fetch_webpage — fetch helpers
-# --------------------------------------------------------------------------- #
-
-
-def _streamed_read(body: bytes):
-    """Build a urlopen-mock read side_effect that serves `body` in chunks."""
-    buf = {"pos": 0}
-
-    def _chunked(n=-1):
-        if n == -1:
-            remaining = body[buf["pos"]:]
-            buf["pos"] = len(body)
-            return remaining
-        chunk = body[buf["pos"]:buf["pos"] + n]
-        buf["pos"] += len(chunk)
-        return chunk
-
-    return _chunked
-
-
-def test_jina_fetch_returns_body_under_cap():
-    from src.tools.web import _jina_fetch
-
-    payload = b"Title: x\nURL Source: https://example.com/\n\nBody here."
-    with patch("src.tools.web.urllib.request.urlopen") as opener:
-        resp = opener.return_value.__enter__.return_value
-        resp.headers = {"Content-Length": str(len(payload)), "Content-Type": "text/markdown"}
-        resp.read.side_effect = _streamed_read(payload)
-        text = _jina_fetch("https://r.jina.ai", "https://example.com/", max_bytes=1024)
-    assert "Body here." in text
-
-
-def test_jina_fetch_content_length_over_cap():
-    from src.tools.web import _jina_fetch
-
-    with patch("src.tools.web.urllib.request.urlopen") as opener:
-        resp = opener.return_value.__enter__.return_value
-        resp.headers = {"Content-Length": "999999"}
-        resp.read.side_effect = _streamed_read(b"x" * 10)
-        with pytest.raises(ValueError, match="MAX_WEB_BYTES"):
-            _jina_fetch("https://r.jina.ai", "https://example.com/", max_bytes=1024)
-
-
-def test_jina_fetch_streamed_over_cap():
-    from src.tools.web import _jina_fetch
-
-    body = b"x" * 2000
-    with patch("src.tools.web.urllib.request.urlopen") as opener:
-        resp = opener.return_value.__enter__.return_value
-        resp.headers = {}  # no Content-Length
-        resp.read.side_effect = _streamed_read(body)
-        with pytest.raises(ValueError, match="MAX_WEB_BYTES"):
-            _jina_fetch("https://r.jina.ai", "https://example.com/", max_bytes=1024)
-
-
-def test_raw_fetch_returns_body_under_cap(monkeypatch):
-    from src.tools.web import _raw_fetch
-
-    body = b"<html><title>Hi</title><body>hello</body></html>"
-
-    # _raw_fetch builds its own opener (with _NoRedirectHandler) and calls opener.open.
-    # Patch urllib.request.build_opener so we intercept that call.
-    fake_opener = MagicMock()
-    cm = fake_opener.open.return_value.__enter__.return_value
-    cm.headers = {"Content-Length": str(len(body)), "Content-Type": "text/html"}
-    cm.read.side_effect = _streamed_read(body)
-    monkeypatch.setattr("src.tools.web.urllib.request.build_opener", lambda *_: fake_opener)
-
-    html = _raw_fetch("https://example.com/", max_bytes=1024)
-    assert "hello" in html
-
-
-def test_raw_fetch_streamed_over_cap(monkeypatch):
-    from src.tools.web import _raw_fetch
-
-    body = b"a" * 4096
-    fake_opener = MagicMock()
-    cm = fake_opener.open.return_value.__enter__.return_value
-    cm.headers = {}
-    cm.read.side_effect = _streamed_read(body)
-    monkeypatch.setattr("src.tools.web.urllib.request.build_opener", lambda *_: fake_opener)
-    with pytest.raises(ValueError, match="MAX_WEB_BYTES"):
-        _raw_fetch("https://example.com/", max_bytes=1024)
-
-
-# --------------------------------------------------------------------------- #
-# fetch_webpage — end-to-end via tool function
-# --------------------------------------------------------------------------- #
-
-
-def _public_dns(monkeypatch):
-    """Route all src.tools.web.socket.getaddrinfo lookups to a public IP."""
-
-    def _public(host, port, family=0, type=0, *args, **kwargs):
-        return [(None, None, None, "", ("93.184.216.34", port))]
-
-    monkeypatch.setattr("src.tools.web.socket.getaddrinfo", _public)
-
-
-def test_default_registry_now_includes_fetch_webpage():
-    names = set(default_registry.names())
-    assert "fetch_webpage" in names
-
-
-def test_fetch_webpage_rejects_http_via_tool(monkeypatch):
-    from src.tools.web import fetch_webpage
-
-    ctx = _ctx()
-    with pytest.raises(ValueError, match="https"):
-        fetch_webpage(ctx, url="http://example.com/")
-
-
-def test_fetch_webpage_jina_happy_path(monkeypatch):
-    from src.tools.web import fetch_webpage
-
-    _public_dns(monkeypatch)
-    jina_body = (
-        b"Title: Example Page\n"
-        b"URL Source: https://example.com/\n"
-        b"Markdown Content:\n"
-        b"# Hello\n\nSee [Docs](https://docs.example.com/) and [Blog](https://blog.example.com/).\n"
-    )
-    ctx = _ctx()
-    with patch("src.tools.web.urllib.request.urlopen") as opener:
-        resp = opener.return_value.__enter__.return_value
-        resp.headers = {"Content-Length": str(len(jina_body))}
-        resp.read.side_effect = _streamed_read(jina_body)
-        out = fetch_webpage(ctx, url="https://example.com/")
-    assert out["source"] == "jina"
-    assert out["title"] == "Example Page"
-    assert "Hello" in out["content"]
-    urls = [link["url"] for link in out["links"]]
-    assert "https://docs.example.com/" in urls
-    assert "https://blog.example.com/" in urls
-    assert out["truncated"] is False
-    assert out["chars"] == len(out["content"])
-
-
-def test_fetch_webpage_falls_back_to_raw_on_jina_5xx(monkeypatch):
-    from src.tools.web import fetch_webpage
-    import urllib.error
-
-    _public_dns(monkeypatch)
-    html_body = (
-        b"<html><head><title>Raw Title</title></head>"
-        b"<body><p>Raw body.</p>"
-        b"<a href='https://docs.example.com/'>Docs</a></body></html>"
-    )
-
-    def jina_fail(*args, **kwargs):
-        raise urllib.error.HTTPError(
-            "https://r.jina.ai/https://example.com/", 500, "boom", {}, None
-        )
-
-    fake_opener = MagicMock()
-    cm = fake_opener.open.return_value.__enter__.return_value
-    cm.headers = {"Content-Length": str(len(html_body))}
-    cm.read.side_effect = _streamed_read(html_body)
-
-    monkeypatch.setattr("src.tools.web.urllib.request.urlopen", jina_fail)
-    monkeypatch.setattr("src.tools.web.urllib.request.build_opener", lambda *_: fake_opener)
-
-    ctx = _ctx()
-    out = fetch_webpage(ctx, url="https://example.com/")
-    assert out["source"] == "raw"
-    assert out["title"] == "Raw Title"
-    assert "Raw body." in out["content"]
-    assert any(link["url"] == "https://docs.example.com/" for link in out["links"])
-
-
-def test_fetch_webpage_jina_body_over_cap_falls_back_to_raw(monkeypatch):
-    """Jina oversize → fall through to raw fetch instead of raising."""
-    from src.tools.web import fetch_webpage
-
-    _public_dns(monkeypatch)
-    settings = Settings(
-        slack_bot_token="xoxb-test",
-        slack_signing_secret="sig",
-        llm_provider="openai",
-        llm_model="gpt-4o-mini",
-        image_provider="openai",
-        image_model="gpt-image-1",
-        agent_max_steps=3,
-        response_language="ko",
-        dynamodb_table_name="t",
-        aws_region="us-east-1",
-        max_web_bytes=128,
-        max_web_chars=8000,
-        max_web_links=20,
-        jina_reader_base="https://r.jina.ai",
-    )
-    ctx = ToolContext(
-        slack_client=MagicMock(),
-        channel="C1",
-        thread_ts="ts1",
-        event={},
-        settings=settings,
-        llm=MagicMock(),
-    )
-
-    huge_jina = b"x" * 4096
-    raw_html = b"<html><body><p>raw small body</p></body></html>"
-
-    fake_opener = MagicMock()
-    cm = fake_opener.open.return_value.__enter__.return_value
-    cm.headers = {"Content-Length": str(len(raw_html))}
-    cm.read.side_effect = _streamed_read(raw_html)
-
-    with patch("src.tools.web.urllib.request.urlopen") as jina_open:
-        jresp = jina_open.return_value.__enter__.return_value
-        jresp.headers = {}  # no Content-Length → streamed-read path
-        jresp.read.side_effect = _streamed_read(huge_jina)
-        monkeypatch_build = patch("src.tools.web.urllib.request.build_opener", lambda *_: fake_opener)
-        with monkeypatch_build:
-            out = fetch_webpage(ctx, url="https://example.com/")
-    assert out["source"] == "raw"
-    assert "raw small body" in out["content"]
-
-
-def test_fetch_webpage_max_chars_truncates(monkeypatch):
-    from src.tools.web import fetch_webpage
-
-    _public_dns(monkeypatch)
-    long_body = b"Title: T\nURL Source: https://example.com/\nMarkdown Content:\n" + (b"A" * 500)
-    ctx = _ctx()
-    with patch("src.tools.web.urllib.request.urlopen") as opener:
-        resp = opener.return_value.__enter__.return_value
-        resp.headers = {"Content-Length": str(len(long_body))}
-        resp.read.side_effect = _streamed_read(long_body)
-        out = fetch_webpage(ctx, url="https://example.com/", max_chars=200)
-    assert out["truncated"] is True
-    assert out["chars"] == 200
-    assert len(out["content"]) == 200
-
-
-def test_fetch_webpage_both_paths_fail_raises(monkeypatch):
-    from src.tools.web import fetch_webpage
-    import urllib.error
-
-    _public_dns(monkeypatch)
-
-    def jina_500(*args, **kwargs):
-        raise urllib.error.HTTPError(
-            "https://r.jina.ai/...", 500, "jina boom", {}, None
-        )
-
-    def raw_503(*args, **kwargs):
-        raise urllib.error.HTTPError(
-            "https://example.com/", 503, "raw boom", {}, None
-        )
-
-    fake_opener = MagicMock()
-    fake_opener.open.side_effect = raw_503
-
-    monkeypatch.setattr("src.tools.web.urllib.request.urlopen", jina_500)
-    monkeypatch.setattr(
-        "src.tools.web.urllib.request.build_opener", lambda *_: fake_opener
-    )
-
-    ctx = _ctx()
-    with pytest.raises(ValueError, match=r"jina=.*raw="):
-        fetch_webpage(ctx, url="https://example.com/")
-
-
-def test_fetch_webpage_max_links_dedup(monkeypatch):
-    from src.tools.web import fetch_webpage
-
-    _public_dns(monkeypatch)
-    link_section = (
-        "[a](https://a.example/)"
-        "[b](https://b.example/)"
-        "[c](https://c.example/)"
-        "[d](https://d.example/)"
-        "[dup-a](https://a.example/)"
-        "[e](https://e.example/)"
-        "[dup-b](https://b.example/)"
-        "[f](https://f.example/)"
-    )
-    payload = (
-        "Title: T\nURL Source: https://example.com/\nMarkdown Content:\n" + link_section
-    ).encode()
-    ctx = _ctx()
-    with patch("src.tools.web.urllib.request.urlopen") as opener:
-        resp = opener.return_value.__enter__.return_value
-        resp.headers = {"Content-Length": str(len(payload))}
-        resp.read.side_effect = _streamed_read(payload)
-        out = fetch_webpage(ctx, url="https://example.com/", max_links=3)
-    urls = [link["url"] for link in out["links"]]
-    assert urls == [
-        "https://a.example/",
-        "https://b.example/",
-        "https://c.example/",
-    ]
